Move content of `/tests` to `/tests/jest` (#857)
