+++ jest.config.js
@@ -5,13 +5,13 @@ module.exports = {
   ],
   moduleNameMapper: {
     '\\.scss$': 'identity-obj-proxy',
-    '\\.svg$': '<rootDir>/tests/mocks/svgrMock.jsx',
+    '\\.svg$': '<rootDir>/tests/jest/mocks/svgrMock.jsx',
   },
   setupFiles: [
-    '<rootDir>/tests/setupJest.js',
+    '<rootDir>/tests/jest/setupJest.js',
   ],
   setupFilesAfterEnv: [
-    '<rootDir>/tests/setupTestingLibrary.js',
+    '<rootDir>/tests/jest/setupTestingLibrary.js',
   ],
   testEnvironment: '@happy-dom/jest-environment',
   transformIgnorePatterns: [
