+++ src/components/FormLayout/__tests__/FormLayout.test.jsx
@@ -3,8 +3,8 @@ import {
   render,
   within,
 } from '@testing-library/react';
-import { childrenEmptyPropTest } from '../../../../tests/propTests/childrenEmptyPropTest';
-import { idPropTest } from '../../../../tests/propTests/idPropTest';
+import { childrenEmptyPropTest } from '../../../../tests/jest/propTests/childrenEmptyPropTest';
+import { idPropTest } from '../../../../tests/jest/propTests/idPropTest';
 import { Button } from '../../Button';
 import { CheckboxField } from '../../CheckboxField';
 import { InputGroup } from '../../InputGroup';
