+++ src/components/Toolbar/__tests__/ToolbarGroup.test.jsx
@@ -3,10 +3,10 @@ import {
   render,
   within,
 } from '@testing-library/react';
-import { alignPropTest } from '../../../../tests/propTests/alignPropTest';
-import { childrenEmptyPropTest } from '../../../../tests/propTests/childrenEmptyPropTest';
-import { densePropTest } from '../../../../tests/propTests/densePropTest';
-import { noWrapPropTest } from '../../../../tests/propTests/noWrapPropTest';
+import { alignPropTest } from '../../../../tests/jest/propTests/alignPropTest';
+import { childrenEmptyPropTest } from '../../../../tests/jest/propTests/childrenEmptyPropTest';
+import { densePropTest } from '../../../../tests/jest/propTests/densePropTest';
+import { noWrapPropTest } from '../../../../tests/jest/propTests/noWrapPropTest';
 import { ToolbarGroup } from '../ToolbarGroup';
 
 const defaultProps = {
