+++ src/components/Toolbar/__tests__/ToolbarItem.test.jsx
@@ -3,7 +3,7 @@ import {
   render,
   within,
 } from '@testing-library/react';
-import { childrenEmptyPropTest } from '../../../../tests/propTests/childrenEmptyPropTest';
+import { childrenEmptyPropTest } from '../../../../tests/jest/propTests/childrenEmptyPropTest';
 import { ToolbarItem } from '../ToolbarItem';
 
 const defaultProps = {
