+++ tests/jest/mocks/svgrMock.jsx
