+++ tests/jest/propTests/actionColorPropTest.js
