+++ tests/jest/propTests/alignPropTest.js
