+++ tests/jest/propTests/isLabelVisibleTest.js
