+++ tests/jest/propTests/labelPropTest.js
