+++ tests/jest/propTests/neutralColorPropTest.js
