+++ tests/jest/propTests/raisedPropTest.js
