+++ tests/jest/propTests/sizePropTest.js
