+++ tests/jest/propTests/variantPropTest.js
